requirements.txt
Flask
python-dotenv
openai
pymongo
gunicorn
